+++ tests/test_scheduler.py
@@ -27,35 +27,35 @@ from nemo_nowcast import scheduler
 
 @patch('nemo_nowcast.scheduler.logging')
 @patch('nemo_nowcast.scheduler.Config')
-@patch('nemo_nowcast.scheduler.lib.base_arg_parser')
+@patch('nemo_nowcast.scheduler.CommandLineInterface')
 @patch('nemo_nowcast.scheduler._install_signal_handlers')
 @patch('nemo_nowcast.scheduler.run')
 class TestMain:
     """Unit tests for scheduler.main function.
     """
-    def test_commandline_parser(
-        self, m_run, m_ish, m_arg_parser, m_config, m_logging,
+    def test_commandline_interface(
+        self, m_run, m_ish, m_cli, m_config, m_logging,
     ):
         scheduler.main()
-        args, kwargs = m_arg_parser.call_args_list[0]
+        args, kwargs = m_cli.call_args_list[0]
         assert args[0] == 'scheduler'
         assert 'package' in kwargs
         assert 'description' in kwargs
-        m_arg_parser().parse_args.assert_called_once_with()
+        m_cli().build_parser.assert_called_once_with()
 
-    def test_config(
-        self, m_run, m_ish, m_arg_parser, m_config, m_logging,
-    ):
-        m_arg_parser().parse_args.return_value = Mock(
+    def test_cli_parser(self, m_run, m_ish, m_cli, m_config, m_logging,):
+        scheduler.main()
+        m_cli().parser.parse_args.assert_called_once_with()
+
+    def test_config(self, m_run, m_ish, m_cli, m_config, m_logging):
+        m_cli().parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         scheduler.main()
         m_config().load.assert_called_once_with('nowcast.yaml')
 
     @patch('nemo_nowcast.scheduler.logger')
-    def test_logging(
-        self, m_logger, m_run, m_ish, m_arg_parser, m_config, m_logging,
-    ):
-        m_arg_parser().parse_args.return_value = Mock(
+    def test_logging(self, m_logger, m_run, m_ish, m_cli, m_config, m_logging):
+        m_cli().parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         m_config.file = 'nowcast.yaml'
         m_config().load.return_value = {'logging': {}}
@@ -64,18 +64,15 @@ class TestMain:
             m_config().__getitem__())
         assert m_logger.info.call_count == 2
 
-    def test_install_signal_handlers(
-        self, m_run, m_ish, m_arg_parser, m_config, m_logging,
+    def test_install_signal_handlers(self, m_run, m_ish, m_cli, m_config, m_logging,
     ):
-        m_arg_parser().parse_args.return_value = Mock(
+        m_cli().parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         scheduler.main()
         m_ish.assert_called_once_with()
 
-    def test_run(
-        self, m_run, m_ish, m_arg_parser, m_config, m_logging,
-    ):
-        m_arg_parser().parse_args.return_value = Mock(
+    def test_run(self, m_run, m_ish, m_cli, m_config, m_logging):
+        m_cli().parser.parse_args.return_value = Mock(
             config_file='nowcast.yaml')
         scheduler.main()
         m_run.assert_called_once_with(m_config())