+++ CHANGES.rst
@@ -5,6 +5,9 @@ Change Log
 In development
 ==============
 
+* Eliminated lib module by refactoring command-line argument parseing
+  functions into attr.s-decorated nemo_nowcast.cli.CommandLineInterface
+  class that is available in the nemo_nowcat namespace.
 * Refactored system config data structure and lib.load_config() into
   attr.s-decorated nemo_nowcast.config.Config class that is available
   in the nemo_nowcat namespace.